Add an interact/get to grab fortune cookies and drink a potion (kinda)
Added the interact/get command which is where we want to put general easter egg interactions in rooms, like getting fortunes from fortune cookies in the kitchen, and getting a potion in the bar to have some random effect.
Right now this is dislaying for all users instead of just you - might want to change.
Potion doesn't actually change your username as intended.
Items are not locked to the room they're found in.