Fix CSS issue with iframes
It was restricting the Theater embed to width=0.

I looked it up and apparently it's supposed to make it more responsive
on mobile, but I tried doing the mobile mode in chrome dev tools and
manually checking all the iframes and couldn't see any difference.

I am not a frontend expert though.