Explicit null return seems required?
Without this, I hit an issue where LeftView and EnteredView weren't rendering and this made the whole site not load. My guess was that a change made a previously valid roomId become invalid - based on the console error messages, what it wanted in those cases was just an explicit null return.