Fix locked custom emoji having broken images
If the emoji was custom, then when it was trying to show the locked version, it would still match to b.isCustom and try to use an image. But the src won't be able to match anything valid, so it is a broken image.
Switched it so that it relies on the isCustom coming from the props, so that we can make sure it's always false for locked badges.