Fix (hopefully?) janky scroll behavior
Previously it had essentially been hand-calculating offsets for the top
of the scroll element, I basically just removed the offsets; I believe
that it was intended to work with the math of the newly added elements
to track what is visible and what needs to be rendered, but you don't
actually need to offset it by the width of the scroll on the page to do
that I think?

Anyways this keeps the virtualization element intact - spam words and
then open inspector on it and scroll up/down and you'll see the list
being of the same size.