Restore potions, clarify theatre and feedback spaces
Brought the polymorph potion and colour-changing function into new rooms, with an update to the theming.
Renamed the Cineplex to Theatre, so it's easier to direct attendees and less branded.
Renamed the Kiosk so it is clearer it's where you can give site feedback, and added the site feedback notewall.