Issue #843 - Disable Google SSO
Firebase changed how the API worked and I can't figure out how to get
Azure to properly reverse proxy the calls. It's like some sorta CORS but
for browser cookies and anywayways it's apparently remediable but I'll
be dammned if I have the time and energy to fix it, especially since
preview is this Friday.

Anyways we're disabling everything but email and I hope that works lmao