Bandange possible room desyncs
Changes the server to return the roomId on move commands and all chat
messages. If the user has intentionally moved (via the UI or /move) then
there will be a "moved: True" in the body.

Changes the client to read the response and, if the room is different,
to call the server to move to the current client room.

I don't necessarily expect this to fix everything, but hopefully it
fixes some of it...