Expand tabs to spaces in source code, and add .editorconfig
A couple of the source files had mixes of tabs and spaces, though most of the code uses just spaces for indents. In particular, roomData.json was using a 2-space indent and 4-space tabs, which looks ragged in some display contexts, like VS Code where the indent and tab size are usually the same. This commit expands the tabs to spaces to make the code display uniform in all contexts.

Converts src/config.ts to Unix mode line endings, to match the rest of src/*.ts files

Also adds an .editorconfig file to make use of spaces instead of tabs, and our typical indentation size, automatic for editors which support it.